Refactor to feature-based structure
